Tpetra: Add CMake option Tpetra_ASSUME_CUDA_AWARE_MPI (part of #1571)
@trilinos/tpetra Add a CMake option Tpetra_ASSUME_CUDA_AWARE_MPI, with
associated macro TPETRA_ASSUME_CUDA_AWARE_MPI defined in
TpetraCore_config.h.  If the CMake option is ON, Tpetra may assume
that the MPI implementation it uses is CUDA aware.  See #1571 for
discussion, and #1088 for an application.

The option currently defaults to OFF.  #1571 requires that we actually
have a test for this option, at least for OpenMPI, so we haven't
finished the issue yet.