Teuchos::TypeNameTraits::demangleName: Fix memory leak & warning
@trilinos/teuchos Fix a potential memory leak in
Teuchos::TypeNameTraits::demangleName.  In a debug build, the function
checks whether abi::__cxa_demangle returned a nonzero status or a NULL
char*.  It could be possible for the function to return a nonzero
status _and_ a nonnull char*.  In that case, the code was failing to
free the char* before throwing an exception.  This commit fixes that.
It also fixes an "unreachable code" warning.