Tpetra: Fix #1623
@trilinos/tpetra Tpetra::Details::FixedHashTable used to inherit from
Teuchos::Describable.  This was a problem, because FixedHashTable
needs to be constructed and destructed on device, but Describable's
methods (e.g., its destructor) were not marked as __device__
functions.  This commit fixes that, and thus fixes #1623, by making
FixedHashTable no longer inherit from Describable.

FixedHashTable is an implementation detail of Tpetra, since it lives
in the Tpetra::Details namespace.  Thus, this change does not
technically break backwards compatibility.  I left in the
implementations of the Describable methods (describe() and
description()); this will hopefully minimize backwards compatibility
issues.