Tpetra: Add unit test for CUDA-aware MPI
@trilinos/tpetra See #1571 and #1088 to learn what it means for an MPI
implementation to be "CUDA aware," and why this matters for
performance.

This commit adds a test for whether the MPI implementation that
Trilinos uses is actually CUDA aware, if it claims to be CUDA aware,
or if the user claims that MPI is CUDA aware.

The test will only build if CUDA is enabled.  If you want to exercise
this test, you must build Tpetra with CUDA enabled, and set the
environment variable TPETRA_ASSUME_CUDA_AWARE_MPI to some true value
(e.g., "1" or "TRUE").  If you set the environment variable to some
false value (e.g., "0" or "FALSE"), the test will run but will pass
trivially.  This means that you may control the test's behavior at run
time.