Intrepid2: Added tests for DeRham complex properties for Hex, Tri, and Quad. (#1586)
For Hexahedrons we test that the grad of an HGrad function is
in the associated HCurl space, the curl of an HCurl function is
in the associated HDiv space, and the div of an HDiv function is
in the associated L2 space:

HGRAD_Cn--(grad)-->HCURL_In--(curl)-->HDIV_In--(div)-->L2_C(n-1)

For 2D topologies (Triangles and Quads) we test that the grad
of an HGrad function is in HCurl, the curl (or perpendicular grad)
of an HGrad function is in HDiv, and both the curl of an HCurl
function and div of an HDiv function are in L2:

HGRAD_Cn--(grad)-->HCURL_In--(curl)-->L2_C(n-1)
HGRAD_Cn--(curl)-->HDIV_In--(div)-->L2_C(n-1)

In all cases we check that the curl of the grad of an HGrad
function is zero and the div of the curl of an HCurl (HGrad in 2D)
function is zero.

Tested for orders up to MaxOrder=6.