Ifpack2: Create BlockTriDiContainer (issue #1350).
BlockTriDiContainer implements a large-block-Jacobi, small-block tridiagonal
preconditioner using KokkosBatched. The preconditioner is the stationary
iteration
    x = (1 - a) x + a D"(b - R x),
where a is a damping factor and the matrix is A = D + R. D contains the
tridiagonal matrices. BlockTriDiContainer computes the iteration in exactly this
form; in particular the matrix-vector product is R x, not A x.

It is also straightforward and efficient to use this class for simple
small-block Jacobi. In that case, D contains just the block diagonals.

It is a Container and so works with BlockRelaxation.

It also exposes a useful interface for efficient standalone use. This interface
supports, in particular, norm-based termination of the stationary iteration.

KokkosBatched implements SIMD-batched operations on which BlockTriDiContainer
builds.

Build/Test Cases Summary
Enabled Packages: Ifpack2
Disabled Packages: PyTrilinos,Claps,TriKota
Enabled all Forward Packages
0) MPI_RELEASE_DEBUG_SHARED_PT => passed: passed=740,notpassed=0 (233.76 min)
Other local commits for this build/test group: 3380bb1