Tpetra: Add FAQ entry explaining Stokhos ETI for Tpetra classes
@trilinos/tpetra Tpetra's FAQ (frequently asked questions) file lives
in Trilinos/packages/tpetra/doc/FAQ.txt.  It is a plain text file,
human-readable, with (Emacs) org-mode mark-up.

The new FAQ entry explains how to change Stokhos, if you add explicit
template instantiation (ETI) for a class or function in Tpetra that
depends on a Scalar (e.g., CrsMatrix, MultiVector) or Packet (e.g.,
DistObject) template parameter.  You only need to do this if doing ETI
for your new class or function, or if you add ETI to an existing class
or function for which Tpetra did not do ETI before.