Zoltan2: MJ scaling issue #1538.
1- mj_premigration_option is updated.
Now any positive number behaves the same way for this option.
Instead of modular arithmetic, the migration will happens based on the rank ranges.
This will more likely to preserve the order of the coordinates, so the partitioning
changes because of tie-breaking should be minimized.

2-Added a new option mj_premigration_coordinate_count: that decides how many coordinates
each rank will have at the end of premigration. The number of ranks are chosen based on this math.
This parameter defaults to 32000, which should work fine. Users do not need to provide this
parameter for now.

Build/Test Cases Summary
Enabled Packages: Zoltan2
Disabled Packages: PyTrilinos,Claps,TriKota
Enabled all Forward Packages
0) MPI_RELEASE_DEBUG_SHARED_PT => passed: passed=841,notpassed=0 (51.81 min)