Stop populating the SIERRA/ dir with configured headers
After fixing fixing an misspelled TriBITS argument, this was again populating
the SIERRA/ directory with configured headers.  The current SIERRA Trilinos
integration process does not use this anymore so there is no reason to keep
populating these headers.  Therefore, this has been commented out (and the
misspelling was fixed).

This was triggered by change in TriBITSPub/TriBITS#200 for catching unparsed
arguments.