Tpetra: Partial fix for #628
@trilinos/tpetra Tpetra::CrsGraph::makeColMap needs process ranks
(PIDs) of remote GIDs (those GIDs in each process that are in the
column Map on that process, but not in the domain Map on that
process), in order to sort and group GIDs by PID.  CrsGraph used to
throw away this information, then recompute it in
CrsGraph::makeImportExport, via a Map::getRemoteIndexList call that
may do multiple rounds of MPI communication.  This commit lets
makeColMap share that information with makeImportExport, in order to
avoid those rounds of MPI communication.

This is a partial fix for #628.  A full fix would reuse other
information, like the remote GIDs themselves, and their corresponding
LIDs.