Stokhos,Tpetra: PackTraits::packValue now takes const char[] as input
@trilinos/tpetra There's no need for most PackTraits methods to take
Kokkos::View.  This tends to cause build errors, in confusing
buffer_device_type vs. device_type.  More importantly, it adds
unnecessary complexity to the interface, and may slow down the code.
This commit continues the process of de-View-ing PackTraits, by
changing packValue to take a `char[]` for output, instead of a
Kokkos::View.