MueLu/HHG by Max: move creation of Xpetra::MatrixSplitting to driver
We're moving towards a driver, that allows for result testing. So,
we need to be able to read the matrix in the driver, not inside
any HHG-specific class.

Now, we moved the creation of Xpetra::MatrixSplitting to the driver.
Xpetra::MatrixSplitting reads the matrix from a given file and stores
the original as well as the splitted matrix. Both can be accessed using
appropriate routines.

This follows up on commit bbd6f5dc29ecb33e8f6b6211f6a9b05c280b21a7