Framework: update release files to use SEMS env
The machine the releases have been done on is past its end of life
and does not have the compiler we have determined is our minimum so
it is necessary to move off of this machine. The scripts we have
for releases were written specfically for that machine so they needed
to be updated. Using the SEMS env should make it possible to use
nearly any machine in the future.