Tpetra::CrsMatrix: Add test for Albany issue; it passes (Serial, Cuda)
@trilinos/tpetra I added a test for CrsMatrix doExport that exercises
use cases likely to be found in the following Albany issue:

https://github.com/gahansen/Albany/issues/182

This test may also be relevant to #1752.

The test does a doExport from a fill-complete CrsMatrix with
overlapping row Map, to the following cases of CrsMatrix:

  - DynamicProfile, locally indexed
  - StaticProfile, locally indexed
  - Constant fill-complete CrsGraph, locally indexed

I specifically wrote the test so that the target matrix get some
entries in each row from one process, other entries in the same row
from another process, and other entries from both processes (thus
exercising the Tpetra::ADD CombineMode fully).

The test passes with both Kokkos::Serial and Kokkos::Cuda back-ends.
The CUDA build uses CUDA 8.0.44 and GCC 5.3.0, with a Kepler GPU and
Intel host CPU.  I'll need to try it with a Pascal GPU and IBM POWER8
host CPU to exercise the Albany issue mentioned above.  @ikalash
kindly sent me the configuration for testing on that platform.

See the test documentation (comments in the test .cpp file itself) to
see why the test does not yet exercise the case of a globally indexed
target.