Belos::GmresPolySolMgr: Fix #1720
@trilinos/belos The build warning says that the `polyConverged`
variable in `Belos::GmresPolySolMgr` was set but not used.  It seems
like the variable really should have been used in the code.  However,
it's been literally over four years (note my comment on line 1033 of
the file, dated "27 Apr 2013") and nobody has had time to investigate,
as far as I know.  This commit just comments out the variable wherever
it appears.