PyTrilinos: Update docstring scripts to work with Python 2/3
These scripts were missed in the initial push for upgrading to Python
3, because they used the shebang

    #! /usr/bin/env python

and on my system, when I tested, this used Python 2. Reported in
with Komplex and PyTrilinos enabled, there is an error processing the
Komplex doxygen comments. Once that is resolved, I can push a final
commit for this and close the issue.