Adding OutputStreamTraits to Anasazi solver framework
The current OutputManager restricts printing to processor zero, as
well as the output stream, and calls the MPI_COMM_WORLD by default.
All of these have been identified as issues by users.

As suggested by Mark Hoemmen, Anasazi solvers now have an
OutputStreamTraits class to get a Teuchos::FancyOStream that will
be used for outputting solver status and timing details.  For flexibility,
this formatted output stream can be obtained by the solver in several
different ways.  It can either be passed in through the parameter
list using the parameter "Output Stream" or created through the
OutputStreamTraits for each linear algebra adapter.  The default
behavior for the OutputStreamTraits class is to create a FancyOStream
that output on processor 0.  There is an "Output Processor" parameter
that will be passed to the OutputStreamTraits::getOutputStream() method
in the event that the user wants additional control of the output
processor.