Tpetra: Potential fix for #1860
@trilinos/tpetra The "Albany182" test now passes on CUDA, with
TPETRA_ASSUME_CUDA_AWARE_MPI either TRUE or FALSE.  This may indicate
that we have fixed #1860, but I'll need to test in Albany to make
sure.

Here is a link to the Albany issue:

https://github.com/gahansen/Albany/issues/182

Note that the issue was closed in Albany, because Albany has an easy
work-around (set the environment variable
TPETRA_ASSUME_CUDA_AWARE_MPI=TRUE).  #1860 is the corresponding
Trilinos issue.