Tpetra: Remove old test that is no longer relevant
@trilinos/tpetra The test includes LOTS of header files, but does
nothing.  Its original purpose was to test whether Tpetra objects were
declared in the Tpetra namespace, and not in the global namespace.
That may have been an issue back when Trilinos had to support
compilers that had iffy support for namespaces, but it hasn't been of
concern since I started working on Trilinos.  Furthermore, this gets
much better (and more meaningful) coverage in Tpetra's other unit
tests, that actually test the classes in question.  Any time one of
those other tests do "using Tpetra::CrsGraph;" (for example), they
test that CrsGraph is declared in the Tpetra namespace.

It may still be of interest to test that CrsGraph (and other Tpetra
classes) don't get pulled into the global namespace.  However, if we
even want to start testing that again, we should do so in the unit
tests for each Tpetra class, rather than grouping them all into a
giant file.