PyTrilinos: Improved error messages from test and example scripts
When users try to run example and test scripts before installing
PyTrilinos, they need to use the '-t' or '--testharness' command-line
arguments to tell the scripts to search the local build directory for
the PyTrilinos package. This update prints a message to that effect,
since this requirement is not readily obvious.