ShyLU: Add missing TRIBITS_PACKAGE_DEF() calls (TriBITSPub/TriBITS#200)
This changes are needed in order to fix the configure of these ShyLU packages
to work again after merging in new version of TriBITS that carefully checks
for mistakes like this (see TriBITSPub/TriBITS#200 and
TriBITSPub/TriBITS#232).

It must be that TRIBITS_PACKAGE_DEF() does not need to be called.  I am not
shy how or why that works but to be safe, we need to enforce that so we only
need to test that one pattern in TriBITS.

Build/Test Cases Summary
Enabled Packages: ShyLU_DD, ShyLU_Node
Disabled Packages: PyTrilinos,Claps,TriKota
0) MPI_RELEASE_DEBUG_SHARED_PT => Test case MPI_RELEASE_DEBUG_SHARED_PT was not run! => Does not affect push readiness! (-1.00 min)
1) MPI_RELEASE_DEBUG_SHARED_ST => passed: passed=2,notpassed=0 (0.02 min)
Other local commits for this build/test group: eb466a8