Teuchos: Remove climits include from Teuchos_ConfigDefs.hpp
@trilinos/teuchos Remove <climits> include from
Teuchos_ConfigDefs.hpp.  Fix Teuchos_OrdinalTraits.hpp to use
std::numeric_limits instead of the LLONG_MAX constant in <climits>.

This commit may break the build of downstream applications that depend
on header files implicitly included via Teuchos_ConfigDefs.hpp.  This
is why I will do this a few includes at a time, to facilitate
bisection.  This is worth the risk because it will improve build times
and encapsulation.