Teuchos: Remove fstream include from Teuchos_ConfigDefs.hpp
@trilinos/teuchos Remove <fstream> include from
Teuchos_ConfigDefs.hpp.  Add "#include <fstream>" to packages' files
that need it.  Affected packages:

  - Domi
  - EpetraExt
  - ML (code that depends on Epetra and Teuchos)
  - NOX
  - ROL (SEE NOTE BELOW)
  - Rythmos
  - Stokhos
  - Tempus
  - Teuchos
  - Zoltan2

This commit may break the build of downstream applications that depend
on header files implicitly included via Teuchos_ConfigDefs.hpp.  This
is why I will do this a few includes at a time, to facilitate
bisection.  This is worth the risk because it will improve build times
and encapsulation.

NOTE: ROL lives in a separate repository, and is periodically
snapshotted into Trilinos.  This commit's changes to ROL may be
clobbered by future snapshots.  This Trilinos issue shows the
necessary changes to ROL, which the ROL developers must do in ROL's
respository:

https://github.com/trilinos/Trilinos/issues/1994