Modify the getNodalFieldSolution function in FEI.
Previous to this change, the nodal loop would continue if the node was
a shared node and was not attached to the local element. This continue
was causing issues in parallel when the nodal data was being synched across
processor after a FETI solve when the run involved element death.
This created many issues and support tickets we have been working through for a
long time.

This was a great team effort with Alan, Kendall, Mike, and Patrick.
Alan and Kendall will work on pushing this change to Trilinos.