Teuchos: Remove typeinfo include from Teuchos_ConfigDefs.hpp
@trilinos/teuchos Remove <typeinfo> include from Teuchos_ConfigDefs.hpp.
Add "#include <typeinfo>" to packages' files that need it.  Affected
packages: Teuchos.

This commit may break the build of downstream applications that depend
on header files implicitly included via Teuchos_ConfigDefs.hpp.  This
is why I will do this a few includes at a time, to facilitate
bisection.  This is worth the risk because it will improve build times
and encapsulation.