Add script to drive snapshotting of TriBITS (TriBITSPub/TriBITS#242)
I decided to create a script to further simplify the process of snapshotting
TriBITS into Trilins.  This makes sure the snapshotting is always run from the
right directories.

But the main reason that I added this script was to make sure that the new
option --clean-ignored-files-orig-dir is used.  That ensures that we don't
accidentally copy over locally ingored files like happened with the file:

  cmake/tribits/doc/developers_guide/TribitsBuildReference.html

as described in TriBITSPub/TriBITS#242.