Teuchos: move SetScientific to Core
This functionality to set the precision
of a C++ stream such no losses are
incurred for floating point values,
and a convenient scope guard for it,
seem general enough to be take out of
the MatrixMarket sub-package.
Anecdotally, I have an application that
wants to use this, but it seems excessive
to require it to depend on MatrixMarket.

Although I left a backward-compatibility
header in the code, I did take the liberty
of updating the usage of this class in the
following packages:
 - Anasazi
 - Belos
 - Tpetra