Fix compilation error when Triutils is not enabled, but Epetra is enabled
When the utilities for creating a linear problem from a file was moved into
the belos/epetra/src directory, there should have been an optional dependence
added for Triutils for the source files.  To build these utilities, the path
to the Triutils headers needs to be on the link line.  It is there for the examples
and tests, but not the source files since the optional dependence was not stated.
Now these Epetra utilities will build cleanly.