ExplicitModelEvaluator: fix a little weird SEGFAUL bug
Per Roger's advice: When using explicit stepper, if the response are evaluated before a
single step is taken (and f is not set in OutArgs), then `scrap_f` is
never allocated and `Thyra::Vt_S` attempts to scale a null pointer and
and a SEGFAULT occurs. Just checking that `f` is not null before
attempting to scale it.