Skip TriBITS setting C++11 and OpenMP flags when Kokkos flags are used (#1400)
This required a small change to TriBITS to allow the OpenMP flags to be simply
skipped.

I also adde an if statement to avoid adding any of the Kokkos flags or
changing any of this logic when KOKKOS_ARCH is equal to "None".  This is
likely important to maintain backward compatiblity.