Disable test that was enabled when Scotch TPL was enabled (#2051, #2052)
This test was never run in any automated builds of Trilinos that posted to
CDash.  It looks like it was only enabled after enabling the Scotch TPL.

This test is only being disabled in the standard CI build.  It is not being
disabled for any other build of Trilinos.  It is necessary to temporarily
disable this test so that we can turn on testing with Scotch with the rest of
Trilinos (which is needed to reproduce the build failure reported in #1400).