MueLu/HHG: add alternative ways to compute coarse col map and prolongator
Ray implemented an alternative way to compute the coarse level column map and
the prolongator. It is based on integer artithmetic and is supposed to mimic
a realistic scenario, i.e. data that is usually used to construct a geometric
prolongator.

Note: A mysterious seg fault occurs in the 'Terminate' command. We don't care.