Teuchos: stop throwing from destructors
This is the foundation for fixing #1303.
First, the macro TEUCHOS_TEST_FOR_TERMINATION
is introduced, which calls std::terminate()
instead of throwing an exception.

I think std::terminate() is the most appropriate
exit call here, because it is exactly what the
compiler would have done anyway. In particular,
the warnings we were getting said
"warning: throw will always call terminate() [-Wterminate]"
As such, if we explicitly call std::terminate ourselves,
we are preserving the existing behavior while silencing
the warning.

TEUCHOS_TEST_FOR_TERMINATION takes stream message code,
and that message will be sent to std::cerr.

This commit also replaces all calls in Teuchos to
TEUCHOS_TEST_FOR_EXCEPTION (and similar macros) from
within destructors with TEUCHOS_TEST_FOR_TERMINATION.

The most important change is the one in ~RCPNodeTmpl()
in Teuchos_RCPNode.hpp which is guarded by TEUCHOS_DEBUG.
When TEUCHOS_DEBUG is defined, this instance alone emitted
its warning at least 136702 times in one user's reported build
of Trilinos, due to the ubuquitous use of RCP in Trilinos.
Given that one warning and its associated notes are close to 1KB
of text, this dominated and bloated all other build output
for a total of around 100MB of text output from the build.