Teuchos: try to silence RCPNodeInfo() warning
It seems certain compilers are smart enough
to realize that no code will ever call
RCPNodeInfo's default constructor, and it
is therefore pointless to define it.
To make this clear, I've changed its definition
to be "= delete;" (C++11 !) and changed the
one place where RCPNodeInfo is constructed to
use emplace_hint, which really isn't too
critical but may improve performance a bit.