Add env script for SEMS ATDM GCC 7.2 env (#2028)
This is a drop-in replacement for load_sems_dev_env.sh to be able to configure
and build Trilinos, but with a custom set of modules.  Just soruce this and
then you can use the SEMSEnv.cmake module.

NOTE: Had to switch from CMake 3.10.0 to CMake 3.5.1 to avoid FIND_LIBRARY()
defect (#2028).

We will need to get Kitware to fix that defect otherwise will not be able to
use the all-at-once configure, build, test, and submit to CDash :-(

Or, we can just hack the SEMSEnv.cmake file to avoid the find like in commit
0960c822a186216ef394fee5b8e9efce50d7585c.