Suppress GlobalMPISession startup banner (#1303, #2089)
The test TeuchosCore_testTeuchosTestForTermination_1_MPI_4 failed today on one
of the test machines due to mangling of the ouptut from the GlobalMPISession
starup banner on one of the other proces and the error message that gets
printed.  This resulted in the regex failing.

Here I just added --teuchos-suppress-startup-banner that suppresses this.

Build/Test Cases Summary
Enabled Packages: TeuchosCore
Disabled Packages: PyTrilinos,Claps,TriKota
0) MPI_RELEASE_DEBUG_SHARED_PT => passed: passed=59,notpassed=0 (1.01 min)