Belos: Modifications to CG
Basically, we allow CGIter as a bypass for the single RHS case in
PseudoBlockCGSolMgr (much as it is done in BlockCGSolMgr).  This involves
pushing stubs for the condition number estimation up to CGIteration (though
they don't do anything).

Basically, this allows us to bypass a chunk of CPU data manipulation time when
running w/ Tpetra on CUDA.