MueLu: fixing issue with multiple dof per node repartitioning for Blocked system
The problem was that a rowMap for the problem was extracted from the
Coordinates vector which has no knowledge of dofPerNode. Replaced that with
 an if statement that checks whether the matrix is blocked in which case
we extract the fullMap, otherwise we extract the rowMap directly.
Also clean up a bit and removed some &* in favor of getRawPtr to extract
c++ pointers from RCP objects.