Tpetra,Ifpack2: Fix #2290 (Purge 'classic' template parameter)
@trilinos/tpetra @trilinos/ifpack2

WARNING: This will break code (e.g., Nalu) that does forward
declarations of Tpetra classes.  Use of the 'classic' template
parameter was already deprecated in Trilinos 12.0, and we need to get
rid of this in order to move ahead with FY18 Tpetra plans.

I only made tiny changes to Ifpack2.  Almost all the changes were in
Tpetra.