Update to clealry show why the test passes or fails (#2247)
Before, if any of the three criteria failed, it would jsut print "Test
failed".  But now it prints why it failed with details.

This test currently fails with Intel compilers (see #2247) but at least now it
shows you why (which was not clear at all before).

I also made usage of the default FancyOStream to avoid logic about what
process you are on for when you should be printing or not.  That is the best
way to handle parallel output and better test output control.