Change how jobs are run on white/ride (TRIL-171)
The "#PBS" options in the script bsub-ctest-job.sh were not being read at all.

This has been updated to excludity using the rhel7F processors since that is
what actually runs the Kokkos tests sucessfully (with KOKKOS_ARCH=Kepler37).

I also updated to use CMake 3.10.1 + Ninja now that CMake 3.10.1 was installed
on ride and white.  This seems to work fine now with the CUDA build.