Panzer: add (optional) prefix to response field name
To help identify the response type on the evaluated field to help
prevent a bug when `Functional` and `ExtremeValue` responses are use at
the same time on the same field.

Before, a single node was shared and depending on the order the response
sublist for two types were added, the result would be different (correct
for the first response but wrong for the second one).

This should help fix that.