Require exclude listing of SEMSDevEnv.cmake, remove load_ci_sems_dev_env.sh (#482)
The logic that resulted in the SEMSDevEnv.cmake file getting picked up was a
little too magical.  Therefore, this change is to require that you explictily
list cmake/std/sems/SEMSDevEnv.cmake in the Trilinos_CONFIGURE_OPTIONS_FILE
argument.  This also got rid of the StdDevEnvs.cmake file and gets rid of that
appraoch.  Having to list a single file in Trilinos_CONFIGURE_OPTIONS_FILE is
not that big of a deal and it is explicit with no magic.

I also got rid of the script load_ci_sems_dev_env.sh because it just soruces
load_sems_env.sh with no arguments by default anyway and we will maintain that
going forward.  This just reduces clutter.