Allow enable of Panzer examples on main builds (#2319)
Turns out that some of these Panzer examples test behavior that EMPIRE needs.
Therefore, we need to get them working.  Therefore, to help get these fixed,
and since the rest of the cuda build is not clean yet, we need to turn these
back on.

Note that this should not enable the Panzer examples for the special "-panzer"
builds since the CTest -S driver script will explicitly disable the Panzer
examples for those builds.

Build and test results on 'shiller' show below (the build passes but there are
still some failing test).  These builds are not promoted to the "ATDM"
Group/Track yet so this will not spam any one with CDash error emails.

Enabled Packages: Panzer

Build test results:
-------------------
1) cuda-opt => FAILED: passed=150,notpassed=3 => Not ready to push! (55.40 min)
2) cuda-debug => FAILED: passed=151,notpassed=2 => Not ready to push! (66.06 min)