Disable timing-out Panzer tests after reducing to ctest -j8 (#2446)
After reducing from ctest -j16 to -j8, less of these Panzer tests are timing
out at 10 minutes.  The remaining tests that are timing out on at least one
reamining build of Panzer on 'ride' and 'shiller' are:

* PanzerAdaptersSTK_main_driver_energy-ss-loca-eigenvalue
* PanzerAdaptersSTK_MixedPoissonExample-ConvTest-Hex-Order-3

And I disabled these tests only for the builds and systems where they timed
out when I ran them locally.

But there are still some failing Panzer test on 'ride'/'white' for the build
gnu-opt-openmp but I think those may be related to the compiler defect on this
system documented in #1208.