Tpetra: Requested Updates to FEM Assembly (#2457)
* Tpetra: Issues from PR #2443

- [x] Add --with-StaticProfile argument
- [x] Changed args to specify only 'type' of FEM Assembly
  - `--with-insert-global-indices`
  - `--with-local-element-loop`
  - `--with-total-element-loop`
- [x] Nuanced header guards more by adding `TPETRAEXAMPLES_` prefix.
- [x] Allowing only one type to be run with each execution
- [x] Using `const comm_ptr_t&` in parameters to executors for each type
      rather than non-const version.

Relates to Issue #2453

* Tpetra: FEM Assembly Updates for #2453

- [x] Remove `Tpetra_getDefaultPlatform` from includes
- [x] Replace `Tpetra::DefaultPlatform::getDefaultPlatform().getComm()`
with `Tpetra::getDefaultComm()` in `fem_assembly_main.cpp`

Relates to Issue #2453

* Tpetra: Call Tpetra::finalize() on premature exit

Moved `readCmdLineOpts()` into a try{}catch(...){} block which sets the
exit status to FAILURE if the command line processing exits with an
error code or if it throws an error.  Catch sets the status to
`EXIT_FAILURE` which will trigger a premature exit with
`Tpetra::finalize()` called.

This should work until #2459 is addressed.