Tempus: Make "IRK 1 Stage Theta Method" the default Start-Up Stepper.
Changing the default Start-Up Stepper from "RK Backward Euler" to
"IRK 1 Stage Theta Method".  The primary drive for this is that
"IRK 1 Stage Theta Method" is second order with its default theta=0.5,
while "RK Backward Euler" is first order.

Also fixed the reference error slope line written to several
*-Error.dat files.  They were writing a first-order reference
errors, instead of the methods actual order.

Also rebaselined a few of the BDF2 tests, because of the change
in the default Start-Up Stepper.