Stokhos:  Allow mean-based PCE preconditioner with double scalar type.
This adds a Tpetra::Operator to wrap a double-valued Tpetra::Operator as
a PCE-valued one.  This is useful for preconditioners such as MueLu
where using UQ::PCE as the scalar type can be problematic (and is
currently broken).  It has some additional cost in that it needs to
permute the data when applying the (multi-) vector but is very simple.
Change PCE unit-tests to use this (for the time-being at least, until
the issues with the PCE scalar type in MueLu can be resolved).