Use default value in ATDM_SET_ATDM_VAR_FROM_ENV_AND_DEFAULT() (TRIL-171)
Turns out that this function was ignoring the passed in default value (the
second argument) and was just always setting 'OFF' as the default.  But it
seems that all of the current usages of this function did set the default
'OFF' so this change is not going to change any ATDM configurations of
Trilinos.

I noticed this while trying to add another var that did not have the default
value of 'OFF' but this will be used in a later commit.