PyTrilinos: Got gen_teuchos_rcp.py script to work using %import
My first stab at this was to %import *.i files, but this didn't work
because it forced Teuchos/RCP.py to try to import every PyTrilinos
module, and it failed trying to import Teuchos. I changed to %import
the header files, which is sufficient to get SWIG to understand the
Trilinos inheritence tree.

Currently tested for Teuchos + Epetra, but nothing more complicated.