Allow Trilinos_TRACK to already be set in env (#2511, TRIL-171)
This will allow me to set 'export Trilinos_TRACK=Specialized' in the Jenkins
job for this build on 'ride' but still let the cron job on 'white' submit to
the "ATDM" group.  I tested this manually and it does what I expect.

This is because some tests are timing out on 'ride' that are not timing out on
'white'.