Tpetra: Fix #2198 (PR #2512)
* Tpetra::Import: Improve documentation for expert constructor
* Tpetra::Import: New constructor (Fix #2198)

@trilinos/tpetra

@trilinos/tpetra Add the new Tpetra::Import constructor requested in
Issue #2198, and a unit test.  The constructor avoids the
getRemoteIndexList call that the usual Import constructor must do,
and thus requires less MPI communication.  I added unit tests.

I also improved documentation for a separate Tpetra::Import "expert"
constructor.  That constructor is not the same as the constructor that
this commit adds to Import.