PyTrilinos: Update Tpetra and Teuchos.RCP to wrap Tpetra::Map directly
The existing wrappers for Tpetra use copied-and-pasted-and-simplified
declarations for various classes to make the SWIG wrapping easier
(they were written at an inopportune time when both new and classic
were being supported). This gets rid of that for Tpetra::Map by
%including Tpetra_Map_decl.hpp directly. This was done in the hopes of
getting the Teuchos.RCP functions to work on Tpetra objects, but it
didn't work. Nevertheless, it is the proper way to wrap a class, so I
am keeping it. The same should be done for MultiVector, Vector, etc...