Belos: Change all default parameters to be constexpr (#2483)
The purpose of these changes is to prepare for premain
registration for the new DII system. The statics were
causing some initialization order issues that could
sometimes seg fault the managers when they were constructed.

Some older clang versions fail on the linkage of
the constexpr when calling the set method for ParameterList.
Using a static_cast seems to work around this though
the cast is simply assignining the constexpr to their
original type. I'm anticipating that would be removed eventually.

Using c++17 resolves this.

Note this PR is being used to test #2483 now and seems to
resolve all the static linkage issues allowing for pre-main
registration of DII to work. So that PR will depend on this one.