Use Ninja for build and atdm-cmake/3.11.1 module (#2462)
Ninja is faster at building.  Why not use it.  And we need some Ninja testing
in PR testing.

Using CMake 3.11.1 allows for all-at-once submits and faster running of ctest
in parallel.  And it allows for using Ninja and TriBITS generates nice dummy
makefiles.

This removes a hack for CMake 3.11.1 that only worked for my machine crf450.
Now this should work every SNL COE machine that mounts SEMS.