Move the mpiexec args before -np not after (#2699) (#2729)
The instructions on the 'serrano' support page say that these arguments should
go before the -np argument, not after the --n arugment.  Hopefully this will
address some of the failures we are seeing as reported in #2699.

Also, I changed the -np argument to --n to match exactly what the 'serrano'
support page says to do.