Change bulid parallel level from 16 to 8 on 'serrano' (TRIL-200, #2699)
With the change to the ctest -S script and checkin-test-atdm.sh script, the
value of ATDM_CONFIG_BUILD_COUNT is now actually used when using Ninja.
Before it was ingored.

Hopefully this will avoid running out of memory when building on 'serrano'.

Also, I tried to set ATDM_CONFIG_CMAKE_JOB_POOL_LINK=2 but that currently
causes a build error.  I have inquired about that in
https://gitlab.kitware.com/snl/project-1/issues/60.