Switch from srun to salloc on hansen/shiller (TRIL-209)
For some reason, when the bash script was run with 'srun', it could not run
the cmake and ninja installed under my home directory:

  /ascldap/users/rabartl/install/hansen-shiller/

But when I switched to 'salloc', it could.

Besides, I think 'salloc' is the correct command to use to allocate an
interactive shell.  The command 'srun' is supposed to be used instead of
'mpiexec'.  I am not sure sure why it was ever used in the first place.