Send CUDA 9.0 builds to "Specialized" CDash Track/Group and remove old var (#2706, #2854)
From a copy-and-paste these were hard-coded to send to the "ATDM" CDash
Track/Group.  But these should not be promoted yet so I had the Jenkins jobs
set CTEST_TEST_TYPE=Experimental which overrides that.  Now these will go to
"Specializezd" by default, unless overridden in the Jenkins job.

I also got rid of 'export ATDM_CONFIG_USE_MAKEFILES=ON' which is a non-op
since the var name was changed to ATDM_CONFIG_USE_NINJA a while ago.