Increase timeout for tests for build Trilinos-atdm-serrano-intel-opt-openmp-panzer (TRIL-200, #2699)
For some reason, the running of the tests on 'serrano' timed out at 15 minutes
for this build on 4/17/2-18 and 5/24/2018.  Most days it completes in less
that 2 minutes but on 5/23/2018 it took 13m 7s in addition to timing out these
two days.  That must have been a system issue but I am increasing the timeout
some to give this a chance to finish on days like that.