Reduce ctest from 16 to 8 for serial GCC builds and fix path to cmake and ninja on hansen/shiller (#2976)
This setup has a problem of having tests run on top of each other using the
same cores with the GCC 4.9.3 serial executables.  I tried passing different
sets of arguments into 'mpiexec' but that just resulted in "There are not
enough slots available in the system" errors for a bunch of tests.  See #2976
for more details.

The only successful solution to ths problem will likely be an extension to
ctest to control process and thread affinity and a close collaboration with
MPI as described in #2422 and:

* https://gitlab.kitware.com/snl/project-1/issues/17#note_422034

I also fixed the path for cmake and ninja in my home directory to point to
/home/rabartl/.  On the login this can be /ascldap/users/rabartl/ but on the
compute nodes, that directory does not exist.  This fixes being able to
configure and build on the compute nodes on hansen/shiller with just using
'srun'.

I also removed the warning about using 'srun' over 'salloc'.  I misunderstood
how 'salloc' and 'srun' work.