Resolve bootsrtap problem with sym-link.
When I tested on the branch I used a different
job for the openMP build so it could use the
common driver script.  The PR to merge this
uses the existing job which looks for the
old common file PullRequestLinuxGCCTestingSettings.cmake
which of course has been eliminated when the
branch was merged in.  I could have changed
the jenkins job to its final state and that
would break everyone elses PR builds until mine
is merged, but adding the link on a temporary basis
seemed less messy. As soon as the new code is
merged and the jenkins job is changed this sym-link
should be removed.