Increase maximum number of iterations to 110 to address issues #2920 and #3007
The random right hand side vector used to perform these tests can sometimes require
a few more iterations of preconditioned CG than normal.  Normal is about 87 and the
maximum iterations is set at 100.  Some random right hand sides require 104 iterations.
It is not an issue with the solver, but the linear system being solved.  Thus, this
commit increases the maximum iterations to 110 to give the preconditioned CG solver
more iterations if the right hand side vector is not as easy to solve for.