Belos: Add Xpetra TsqrAdaptor; aim to fix #2999
@trilinos/belos @trilinos/xpetra

I've added a TsqrAdaptor for the Xpetra::MultiVector specialization of
Belos::MultiVecTraits.  It compiles and should work (though untested
-- there are no tests in Belos for the Xpetra specialization) with the
Tpetra back-end of Xpetra.  For the Epetra back-end, it is a stub that
compiles but throws std::logic_error if one attempts to use it.

I really wish Belos::MultiVecTraits didn't have an Xpetra
specialization.  If I can find the time, I will remove it in favor of
Belos' run-time polymorphism interface (Belos::MultiVec and
Belos::Operator).  However, we're stuck with it for now.

This commit aims to fix #2999.