Update for 'white' upgrade to CUDA 9.2 (TRIL-125)
This loads the developer pack module:

  devpack/20180308/openmpi/2.1.2/gcc/7.2.0/cuda/9.0.176

and the updated moudle:

  netlib/3.8.0/gcc/7.2.0

I took out the logic for yaml-cpp because the updated devopack does not load
any yaml-cpp module and the ATDM Trilinos build does not need it.

I took out the moudle load for binutils/2.27.0 since the updated devpack
module loads and updated binutils/2.30.0 module which should address the build
errors for the stokhos_muelu library in #3069.