MueLu: GeometricInterpolationPFactory, see issue #3179
adding the stokhos files, modifying the input parameters and modifying the
structured poisson example in trilinosCouplings to test the new geometric
interpolation factory. I mainly needed to provide us with coordinates for
the piece-wise linear interpolation. This example can also be used to exert
 the hybrid aggregation capabilities now.

Rewrite of parts of structured aggregation to use CrsGraph for piece-wise
constant aggregation. This eventually allows to write a simplified version
of the piece-wise constant interpolation prolongator as maps are transfered
 from aggregation to prolongator via the CrsGraph.

Added getCoarseNodesData function to global and local lexi orderings.
Removing the short name RealValuedMultiVector in favor of a typedef in each file.
Adding comprehensive unit test for GlobalLexicographicIndexManager.
Improving the Index Manager unit test to perform a reduction at the end
 of the test that checks results on all ranks before reporting success.