Disable failing SEACAS tests on Mutrino (#3183) (#3213)
The executables are built in a parallel build, but they are
really serial so when run there is some extra info that is
output to stderr which ends up messing with the textual comparison
with the expected "gold" output files.  For now disable the tests
until can figure out a better way of running them.

This should address #3183.