Tempus:  Fix strange rcp node-tracking error on Intel builds.
The autotester showed the intel builds of the new optimization interface
test failing because of the rcp node tracking that is turned on in debug
builds.  For some strange reason, the RCP that is inside the
Teuchos::Array inside this test never is deallocated, but only with the
intel compiler.  The simple solution was to use a std::vector instead.