MueLu: kokkos refactor of IndexManager and StructuredAggreagtion
All the functionalities are implemented in IM and controled using an enum instead of having derived classes.
The refactor of the coupled version of StructuredAggregation is not complete and there are no plans to complete it at this point.
Still lacking a refactor of the local lexicographic mesh layout due to the use of standard library functions in the non kokkos implementation.
Structured aggregation will now only perform local or uncoupled aggregation.
This simplifies dramatically the implementation of IndexManager_kokkos,
AggregationStructuredAlgorithm_kokkos and StructuredAggregationFactory_kokkos.
removing test of structured aggregation kokkos refactor with linAlgebra=Epetra
The refactor works and passes all the tests.
I have added a unit_test_kokkos as well as a test in the test/structured folder.