Switch some 'waterman' 'debug' builds to new 'release-debug' builds (#3633)
I kept the existing cuda-9.2-debug-Power9-Volta70 build since there are
currently not any timing out tests in that build and I figured that the CUDA
build was most likey the one a developer would want to run with a debugger.
But I created a new cuda-9.2-release-debug-Power9-Volta70 build so that we can
avoid having to disable slow Kokkos, KokkosKernels, and other tests that run
super slow with -O0.

I changed the build gnu-debug-openmp-Power9-Volta70 to a
gnu-release-debug-openmp-Power9-Volta70 build since I don't think it is as
important to run this build with a debugger and the full 'debug' build and
this build currently has some timing-out tests for Kokkos and KokkosKernals as
described in #3336.  (The new gnu-release-debug-openmp-Power9-Volta70 build
has not have any timeouts.)  If the APP teams tell us they want a full
gnu-debug-openmp-Power9-Volta70 build, then we will add one back and deal with
the timeouts.

NOTE: By having both 'debug' and 'release-debug' builds, we can be free to
disable some slow tests in the full 'debug' build and not loose much runtime
debug checking since these tests will be running in the 'release-debug' build
(with runtime debug checking enabled).  So going forward, if a test times-out
in the 'debug' build but not the 'release-debug' build, then we will just
disable it in the 'debug' build and move on.

I also renamed the 'cuda' builds to 'cuda-9.2' builds since that is what they
are and that matches the Jenkins drive names.